add ProgramLinkConfiguration
